v6.9

+++ pkg/scaler/simple.go
@@ -18,7 +18,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/AliyunContainerService/scaler/go/pkg/config"
-	model2 "github.com/AliyunContainerService/scaler/go/pkg/model"
+	m2 "github.com/AliyunContainerService/scaler/go/pkg/model"
 	platform_client2 "github.com/AliyunContainerService/scaler/go/pkg/platform_client"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -33,21 +33,22 @@ import (
 
 type Simple struct {
 	config         *config.Config
-	MetaData       *model2.Meta
+	MetaData       *m2.Meta
 	platformClient platform_client2.Client
 	mu             sync.Mutex
 	wg             sync.WaitGroup
-	instances      map[string]*model2.Instance
+	instances      map[string]*m2.Instance
 	idleInstance   *list.List
 	Cnt            int64
 	InitDuration   time.Duration
 	Exec           *[5]int
 	OnInit         int32
 	UseTime        *[7200]uint16
-	IdleChan       chan *model2.CreateRet
+	IdleChan       chan *m2.CreateRet
+	Deleted        sync.Map
 }
 
-func New(metaData *model2.Meta, config *config.Config) Scaler {
+func New(metaData *m2.Meta, config *config.Config) Scaler {
 	client, err := platform_client2.New(config.ClientAddr)
 	if err != nil {
 		log.Fatalf("client init with error: %s", err.Error())
@@ -58,19 +59,19 @@ func New(metaData *model2.Meta, config *config.Config) Scaler {
 		platformClient: client,
 		mu:             sync.Mutex{},
 		wg:             sync.WaitGroup{},
-		instances:      make(map[string]*model2.Instance),
+		instances:      make(map[string]*m2.Instance),
 		idleInstance:   list.New(),
 		Exec:           &[5]int{},
 		OnInit:         0,
 		UseTime:        &[7200]uint16{},
-		IdleChan:       make(chan *model2.CreateRet),
+		IdleChan:       make(chan *m2.CreateRet),
 	}
-	log.Printf("New scaler for app: %s is created", metaData.Key)
+	m2.Printf("New scaler for app: %s is created", metaData.Key)
 	scheduler.wg.Add(1)
 	go func() {
 		defer scheduler.wg.Done()
 		scheduler.gcLoop()
-		log.Printf("gc loop for app: %s is stoped", metaData.Key)
+		m2.Printf("gc loop for app: %s is stoped", metaData.Key)
 	}()
 
 	return scheduler
@@ -79,18 +80,25 @@ func New(metaData *model2.Meta, config *config.Config) Scaler {
 func (s *Simple) Assign(ctx context.Context, request *pb.AssignRequest) (*pb.AssignReply, error) {
 	start := time.Now()
 	instanceId := uuid.New().String()
-	defer func() {
-		log.Printf("AssignEnd,%s,%s,%s,cost %dms", request.RequestId, request.MetaData.Key, instanceId, time.Since(start).Milliseconds())
-	}()
-	log.Printf("AssignStart,%s,%s", request.RequestId, request.MetaData.Key)
+	if m2.Dev {
+		defer func() {
+			m2.Printf("AssignEnd,%s,%s,%s,cost %dms", request.RequestId, request.MetaData.Key, instanceId, time.Since(start).Milliseconds())
+		}()
+		m2.Printf("AssignStart,%s,%s", request.RequestId, request.MetaData.Key)
+	}
 	s.mu.Lock()
 	if element := s.idleInstance.Front(); element != nil {
-		instance := element.Value.(*model2.Instance)
+		instance := element.Value.(*m2.Instance)
 		instance.Busy = true
 		s.idleInstance.Remove(element)
 		s.mu.Unlock()
-		log.Printf("AssignReuse,%s,%s,%s reused", request.RequestId, request.MetaData.Key, instance.Id)
+		m2.Printf("AssignReuse,%s,%s,%s reused", request.RequestId, request.MetaData.Key, instance.Id)
 		instanceId = instance.Id
+		if instance.LastIdleTime != instance.LastStart {
+
+			m2.Printf("reuseConsume,%s,%d", s.MetaData.Key, time.Since(instance.LastIdleTime).Milliseconds())
+
+		}
 		instance.LastStart = time.Now()
 		return &pb.AssignReply{
 			Status: pb.Status_Ok,
@@ -103,11 +111,11 @@ func (s *Simple) Assign(ctx context.Context, request *pb.AssignRequest) (*pb.Ass
 		}, nil
 	}
 	//Create new Instance
-	log.Printf("AssignNew,%s,%s", request.RequestId, request.MetaData.Key)
+	m2.Printf("AssignNew,%s,%s", request.RequestId, request.MetaData.Key)
 	creatStart := time.Now()
 	atomic.AddInt32(&s.OnInit, 1)
 	go s.CreateNew(instanceId, request.RequestId, s.MetaData)
-	var instance *model2.Instance
+	var instance *m2.Instance
 	s.mu.Unlock()
 	select {
 	case createRet := <-s.IdleChan:
@@ -117,16 +125,21 @@ func (s *Simple) Assign(ctx context.Context, request *pb.AssignRequest) (*pb.Ass
 		}
 		instance = createRet.Instance
 	}
+	if instance.LastIdleTime != instance.LastStart {
+		m2.Printf("waitConsume,%s,%d", s.MetaData.Key, time.Since(creatStart).Milliseconds())
+	}
 	//add new instance
 	s.mu.Lock()
 	instance.Busy = true
+	if instance.LastIdleTime != instance.LastStart {
+		m2.Printf("reuseConsume,%s,%d", s.MetaData.Key, time.Since(instance.LastIdleTime).Milliseconds())
+	}
 	instance.LastStart = time.Now()
 	s.instances[instance.Id] = instance
 	atomic.AddInt32(&s.OnInit, -1)
 	s.mu.Unlock()
 	s.InitDuration = time.Since(creatStart)
-	log.Printf("AssignCreated,%s,%s,%dms", request.RequestId, instance.Meta.Key, time.Since(creatStart).Milliseconds())
-
+	m2.Printf("AssignCreated,%s,%s,%dms", request.RequestId, instance.Meta.Key, time.Since(creatStart).Milliseconds())
 	return &pb.AssignReply{
 		Status: pb.Status_Ok,
 		Assigment: &pb.Assignment{
@@ -148,10 +161,12 @@ func (s *Simple) Idle(ctx context.Context, request *pb.IdleRequest) (*pb.IdleRep
 	}
 	start := time.Now()
 	instanceId := request.Assigment.InstanceId
-	defer func() {
-		log.Printf("IdleEnd,%s,%s,cost %dus", request.Assigment.RequestId, request.Assigment.MetaKey, time.Since(start).Microseconds())
-	}()
-	log.Printf("Idle, request id: %s", request.Assigment.RequestId)
+	if m2.Dev {
+		defer func() {
+			m2.Printf("IdleEnd,%s,%s,cost %dus", request.Assigment.RequestId, request.Assigment.MetaKey, time.Since(start).Microseconds())
+		}()
+		m2.Printf("Idle, request id: %s", request.Assigment.RequestId)
+	}
 	needDestroy := false
 	slotId := ""
 	if request.Result != nil && request.Result.NeedDestroy != nil && *request.Result.NeedDestroy {
@@ -162,30 +177,32 @@ func (s *Simple) Idle(ctx context.Context, request *pb.IdleRequest) (*pb.IdleRep
 			s.deleteSlot(ctx, request.Assigment.RequestId, slotId, instanceId, request.Assigment.MetaKey, "bad instance")
 		}
 	}()
-	log.Printf("IdleStart,%s,%s,", request.Assigment.RequestId, request.Assigment.MetaKey)
+	m2.Printf("IdleStart,%s,%s,", request.Assigment.RequestId, request.Assigment.MetaKey)
 	s.mu.Lock()
 	if instance := s.instances[instanceId]; instance != nil {
 		slotId = instance.Slot.Id
 		instance.LastIdleTime = time.Now()
 		if needDestroy {
 			s.mu.Unlock()
-			log.Printf("request id %s, instance %s need be destroy", request.Assigment.RequestId, instanceId)
+			m2.Printf("request id %s, instance %s need be destroy", request.Assigment.RequestId, instanceId)
 			return reply, nil
 		}
 
 		if instance.Busy == false {
 			s.mu.Unlock()
-			log.Printf("request id %s, instance %s already freed", request.Assigment.RequestId, instanceId)
+			m2.Printf("request id %s, instance %s already freed", request.Assigment.RequestId, instanceId)
 			return reply, nil
 		}
 		instance.Busy = false
 		s.mu.Unlock()
 		select {
-		case s.IdleChan <- &model2.CreateRet{Instance: instance}:
-			log.Printf("reuseByChan %s,%s", instanceId, s.MetaData.Key)
+		case s.IdleChan <- &m2.CreateRet{Instance: instance}:
+			m2.Printf("reuseByChan %s,%s", instanceId, s.MetaData.Key)
 		default:
-			log.Printf("push2Idle %s,%s", instanceId, s.MetaData.Key)
+			m2.Printf("push2Idle %s,%s", instanceId, s.MetaData.Key)
+			s.mu.Lock()
 			s.idleInstance.PushFront(instance)
+			s.mu.Unlock()
 		}
 		return &pb.IdleReply{
 			Status:       pb.Status_Ok,
@@ -198,15 +215,19 @@ func (s *Simple) Idle(ctx context.Context, request *pb.IdleRequest) (*pb.IdleRep
 }
 
 func (s *Simple) deleteSlot(ctx context.Context, requestId, slotId, instanceId, metaKey, reason string) {
-	log.Printf("start delete Instance %s (Slot: %s) of app: %s", instanceId, slotId, metaKey)
-	if err := s.platformClient.DestroySLot(ctx, requestId, slotId, reason); err != nil {
-		log.Printf("delete Instance %s (Slot: %s) of app: %s failed with: %s", instanceId, slotId, metaKey, err.Error())
+	m2.Printf("start delete Instance %s (Slot: %s) of app: %s", instanceId, slotId, metaKey)
+	if _, ok := s.Deleted.Load(slotId); ok {
+		return
 	}
-	log.Printf("delete Instance %s (Slot: %s) of app: %s succ", instanceId, slotId, metaKey)
+	if err := s.platformClient.DestroySLot(ctx, requestId, slotId, reason); err != nil {
+		m2.Printf("delete Instance %s (Slot: %s) of app: %s failed with: %s", instanceId, slotId, metaKey, err.Error())
+	}
+	s.Deleted.Store(slotId, true)
+	m2.Printf("delete Instance %s (Slot: %s) of app: %s succ", instanceId, slotId, metaKey)
 }
 
 func (s *Simple) gcLoop() {
-	log.Printf("gc loop for app: %s is started", s.MetaData.Key)
+	m2.Printf("gc loop for app: %s is started", s.MetaData.Key)
 	ticker := time.NewTicker(1000 * time.Millisecond)
 	lastGc := time.Now()
 	//lastUsageCheck := time.Now()
@@ -214,18 +235,18 @@ func (s *Simple) gcLoop() {
 		for {
 			s.mu.Lock()
 			avgExec := s.GetAvgExec()
-			log.Printf("GcPreSize %s,%d,%d,%f,%v", s.MetaData.Key, len(s.instances), s.idleInstance.Len(), avgExec, s.Exec)
+			m2.Printf("GcPreSize %s,%d,%d,%f,%v", s.MetaData.Key, len(s.instances), s.idleInstance.Len(), avgExec, s.Exec)
 			if time.Since(lastGc) > s.config.GetGcInterval(float32(s.InitDuration), avgExec) && s.idleInstance.Len() > 0 {
 				lastGc = time.Now()
 				cur := s.idleInstance.Front()
 				for {
 					next := cur.Next()
 					element := cur
-					instance := element.Value.(*model2.Instance)
+					instance := element.Value.(*m2.Instance)
 					idleDuration := time.Now().Sub(instance.LastIdleTime)
-					if idleDuration > s.config.GetIdleDurationBeforeGC(float32(s.InitDuration), avgExec, s.UseTime) {
-						log.Printf("Idle duration: %dms, time excceed configured: %dms", idleDuration.Milliseconds(),
-							s.config.GetIdleDurationBeforeGC(float32(s.InitDuration), avgExec, s.UseTime).Milliseconds())
+					if idleDuration > s.config.GetIdleDurationBeforeGC(float32(s.InitDuration), avgExec, s.UseTime, s.MetaData) {
+						m2.Printf("Idle duration: %dms, time excceed configured: %dms", idleDuration.Milliseconds(),
+							s.config.GetIdleDurationBeforeGC(float32(s.InitDuration), avgExec, s.UseTime, s.MetaData).Milliseconds())
 						//need GC
 						s.idleInstance.Remove(element)
 						delete(s.instances, instance.Id)
@@ -235,6 +256,7 @@ func (s *Simple) gcLoop() {
 							ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
 							defer cancel()
 							s.deleteSlot(ctx, uuid.NewString(), instance.Slot.Id, instance.Id, instance.Meta.Key, reason)
+							m2.Printf("gcConsume,%s,%d", s.MetaData.Key, idleDuration.Milliseconds())
 						}()
 					}
 					if next == nil {
@@ -243,7 +265,7 @@ func (s *Simple) gcLoop() {
 					cur = next
 				}
 			}
-			log.Printf("GcEndSize %s,%d,%d", s.MetaData.Key, len(s.instances), s.idleInstance.Len())
+			m2.Printf("GcEndSize %s,%d,%d", s.MetaData.Key, len(s.instances), s.idleInstance.Len())
 			s.mu.Unlock()
 			break
 		}
@@ -273,7 +295,7 @@ func (s *Simple) GetAvgExec() float32 {
 			}
 		}
 	}
-	log.Printf("avgExecCompute %d,%f,%f", s.Cnt, execHis, cnt)
+	m2.Printf("avgExecCompute %d,%f,%f", s.Cnt, execHis, cnt)
 	return execHis / cnt
 }
 
@@ -286,51 +308,83 @@ func (s *Simple) Stats() Stats {
 	}
 }
 
-func (s *Simple) CreateNew(instanceId string, requestId string, meta *model2.Meta) {
+func (s *Simple) CreateNew(instanceId string, requestId string, meta *m2.Meta) {
 	ctx := context.Background()
-	result := &model2.CreateRet{
+	result := &m2.CreateRet{
 		Instance: nil,
 		Err:      nil,
 		Msg:      "",
 	}
 	//Create new Instance
-	log.Printf("AssignNew,%s,%s", requestId, meta.Key)
-	resourceConfig := model2.SlotResourceConfig{
+	m2.Printf("AssignNew,%s,%s", requestId, meta.Key)
+	resourceConfig := m2.SlotResourceConfig{
 		ResourceConfig: pb.ResourceConfig{
 			MemoryInMegabytes: s.MetaData.MemoryInMb,
 		},
 	}
+	start := time.Now()
 	slot, err := s.platformClient.CreateSlot(ctx, requestId, &resourceConfig)
 	if err != nil {
 		errorMessage := fmt.Sprintf("create slot failed with: %s", err.Error())
-		log.Printf("createFailed:%s", errorMessage)
+		m2.Printf("createFailed:%s", errorMessage)
 		return
 	}
 	if s.OnInit < 1 && s.idleInstance.Len() > 0 {
-		log.Printf("fastDelete:%s", meta.Key)
+		m2.Printf("fastDelete:%s", meta.Key)
+		m2.Printf("fastConsume1,%s,%d", s.MetaData.Key, time.Since(start).Milliseconds())
 		s.deleteSlot(ctx, requestId, slot.Id, instanceId, meta.Key, "fastDelete")
 		return
 	}
-	log.Printf("AssignInit,%s,%s", requestId, meta.Key)
-	rMeta := &model2.Meta{
+	m2.Printf("AssignInit,%s,%s", requestId, meta.Key)
+	initRec := make(chan *m2.Instance)
+	rMeta := &m2.Meta{
 		Meta: pb.Meta{
 			Key:           meta.Key,
 			Runtime:       meta.Runtime,
 			TimeoutInSecs: meta.TimeoutInSecs,
 		},
 	}
-	instance, err := s.platformClient.Init(ctx, requestId, instanceId, slot, rMeta)
+	go s.init(initRec, requestId, instanceId, slot, rMeta, start)
+	going := true
+	for going {
+		select {
+		case instance := <-initRec:
+			if instance != nil {
+				result.Instance = instance
+				select {
+				case s.IdleChan <- result:
+					m2.Printf("pushIdleSuc:%s", meta.Key)
+					m2.Printf("initConsume,%s,%d", s.MetaData.Key, time.Since(start).Milliseconds())
+				case <-time.After(100 * time.Millisecond):
+					s.deleteSlot(ctx, requestId, slot.Id, instanceId, meta.Key, "fastDelete")
+					m2.Printf("fastConsume2,%s,%d", s.MetaData.Key, time.Since(start).Milliseconds())
+				}
+			}
+			going = false
+		case <-time.Tick(50 * time.Millisecond):
+			if s.OnInit < 1 || s.idleInstance.Len() > 0 {
+				m2.Printf("fastConsume4,%s,%d", s.MetaData.Key, time.Since(start).Milliseconds())
+				s.deleteSlot(ctx, requestId, slot.Id, instanceId, meta.Key, "fastDelete4")
+				going = false
+			}
+		case <-time.After(360 * time.Second):
+			going = false
+		}
+	}
+}
+
+func (s *Simple) init(initRec chan *m2.Instance, requestId string, instId string, slot *m2.Slot, rMeta *m2.Meta, start time.Time) {
+	instance, err := s.platformClient.Init(context.Background(), requestId, instId, slot, rMeta)
 	if err != nil {
 		errorMessage := fmt.Sprintf("create instance failed with: %s", err.Error())
-		log.Printf(errorMessage)
+		m2.Printf(errorMessage)
 		return
 	}
-	result.Instance = instance
 	select {
-	case s.IdleChan <- result:
-		log.Printf("pushIdleSuc:%s", meta.Key)
-	case <-time.After(100 * time.Millisecond):
-		log.Printf("fastDelete2:%s", meta.Key)
-		s.deleteSlot(ctx, requestId, slot.Id, instanceId, meta.Key, "fastDelete")
+	case initRec <- instance:
+		m2.Printf("initSendSuccess")
+	default:
+		s.deleteSlot(context.Background(), requestId, slot.Id, instId, rMeta.Key, "fastDelete3")
+		m2.Printf("fastConsume3,%s,%d", s.MetaData.Key, time.Since(start).Milliseconds())
 	}
 }
